printf with -1

+++ ft_printf.h
@@ -5,8 +5,8 @@
 /*                                                    +:+ +:+         +:+     */
 /*   By: mohkhan <mohkhan@student.42abudhabi.ae>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/07/10 14:45:34 by mohkhan           #+#    #+#             */
-/*   Updated: 2024/07/10 14:45:38 by mohkhan          ###   ########.fr       */
+/*   Created: 2024/07/11 15:02:01 by mohkhan           #+#    #+#             */
+/*   Updated: 2024/07/11 15:02:04 by mohkhan          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ int	ft_putstr(char *str);
 int	ft_hexaddr(unsigned long n);
 int	ft_rnputnbr(int n);
 int	ft_unsignint(unsigned int n);
-int	ft_hexsm(unsigned int n);
+int	ft_hexsm(unsigned long n);
 int	ft_hexcap(unsigned int n);
 
 #endif
